requirements.txt
fastapi
uvicorn
python-multipart
SQLAlchemy
sqlalchemy 
python-dotenv
psycopg2-binary
psycopg2